+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### 功能构建
 
+- 优化部分类中部分方法的行为分析行为。
+  - com.dwarfeng.voucher.impl.service.CheckerInfoMaintainServiceImpl。
+  - com.dwarfeng.voucher.impl.service.VoucherCategoryMaintainServiceImpl。
+  - com.dwarfeng.voucher.impl.service.VoucherCategoryVariableMaintainServiceImpl。
+  - com.dwarfeng.voucher.impl.service.VoucherMaintainServiceImpl。
+  - com.dwarfeng.voucher.impl.service.VoucherVariableMaintainServiceImpl。
+
 - SPI 目录结构优化。
   - 将检查机制的 SPI 接口与抽象类提相关代码文件提升至 `sdk` 模块中。
   - 将推送机制的 SPI 接口与抽象类提相关代码文件提升至 `sdk` 模块中。